Update utils.js
Removed old line of code for the return statement that was temporarily commented out and replaced because it was causing issues loading images on certain components like the ProjectCard component. New line of code that replaced it (return '/assets/${path}') works fine, so I'm removing the comment containing the old code permanently.